Restructure project directory / module hierarchy
Also added string conversion to PatternNode and made some minor tweaks.
