+++ code_duplication/src/common/Nodes/PatternNode.py
@@ -53,3 +53,9 @@ class PatternNode:
 
     def __ne__(self, other):
         return not self.__eq__(other)
+
+    def __str__(self):
+        return f"{self.value}(', '.join{[n.origin for n in self.nodes]})"
+
+    def __repr__(self):
+        return self.__str__()
