+++ code_duplication/src/common/primary_algorithm/anti_unification.py
@@ -1,4 +1,4 @@
-from .Nodes.PatternNode import PatternNode
+from ..nodes.PatternNode import PatternNode
 
 
 def anti_unify(list1, list2, index1, index2, worktable):
@@ -8,11 +8,11 @@ def anti_unify(list1, list2, index1, index2, worktable):
     if list1 == list2:  # if true:
         # check if leaves
         if not list1[index1].child_indices and not list2[index2].child_indices:
-            t = PatternNode(list1[index1], list2[index2], list1[index1].weight, list1[index1].value)
+            t = PatternNode(list1[index1], list2[index2], list1[index1].value)
             return t
         else:  # if not leaves:
             # iterate thru node's children, adding them as new children to the new fake node
-            p = PatternNode(list1[index1], list2[index2], list1[index1].weight, list1[index1].value)
+            p = PatternNode(list1[index1], list2[index2], list1[index1].value)
             for c in list1[index1].child_indices:
                 # call function on children to get their subtrees - RECURSION
                 subtree = anti_unify(list1, list2, c, c, worktable)
