+++ code_duplication/src/common/primary_algorithm/pattern_collection.py
