+++ code_duplication/src/common/secondary_algorithm/comparer.py
@@ -1,6 +1,6 @@
-from .repo_cloner import _clone_repo
-from .module_parser import get_modules_from_dir
-from .benchmark import time_snap
+from ..preprocessing.repo_cloner import _clone_repo
+from ..preprocessing.module_parser import get_modules_from_dir
+from ..utils.benchmark import time_snap
 
 # Minimum combined weight of two nodes required for comparison.
 _MIN_WEIGHT = 80
